Split out the implicit resolver into its own module
